requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml>=4.9.0
Pillow>=11.0.0
flask==2.3.3
flask-session==0.5.0
requests-oauthlib==1.3.1
cryptography>=41.0.0
